+++ example/apps/functions/functions-cache.ts
@@ -7,12 +7,23 @@ import type {
 import axios from 'axios';
 
 const baseURL = 'https://api.jsonbin.io/v3/b';
-const binId = '632f4a58a1610e63863762f0';
+
+const getBinId = (flavor: string): string => {
+	switch (flavor) {
+		case 'production':
+			return '632f4a58a1610e63863362f0';
+		case 'development':
+			return '632f4a58a1610e63863762f0';
+		default:
+			throw new Error(`Unknown flavor: ${flavor}`);
+	}
+};
 
 const jsonbinMasterKey =
 	'$2b$10$SZkinpMz0Faiv/pl4/Nt7OdyYsnHj9p/unvhZUmjNQyzWcj6hIT.m';
 
-export const fetch: FunctionsCacheFetch = async () => {
+export const fetch: FunctionsCacheFetch = async ({ flavor }) => {
+	const binId = getBinId(flavor);
 	const response = await axios.get<FunctionsCache>(
 		`${baseURL}/${binId}/latest`,
 		{
@@ -25,13 +36,18 @@ export const fetch: FunctionsCacheFetch = async () => {
 	return response.data;
 };
 
-export const update: FunctionsCacheUpdate = async (newFunctionsCache) => {
-	const oldFunctionsCache = await fetch();
+export const update: FunctionsCacheUpdate = async ({
+	flavor,
+	newFunctionsCache,
+}) => {
+	const oldFunctionsCache = await fetch({ flavor });
 
 	const mergedFunctionsCache = {
 		...oldFunctionsCache,
 		...newFunctionsCache,
 	};
+
+	const binId = getBinId(flavor);
 
 	await axios.put(`${baseURL}/${binId}`, mergedFunctionsCache, {
 		headers: {