+++ example/apps/functions/script-config.development.ts
@@ -1,0 +1,6 @@
+// Set the environment variables for the development environment for scripting
+process.env.FIREBASE_SERVICE_ACCOUNT = JSON.stringify({
+	clientEmail: '',
+	privateKey: '-----BEGIN PRIVATE KEY-----',
+	projectId: '',
+});