+++ example/apps/functions/scripts/create-user.ts
@@ -1,13 +1,9 @@
 import { firestore } from '$configs/firestore';
+import type { ScriptFunction } from 'nx-cloud-functions-deployer';
 
-export default async () => {
-	console.log('flavor', process.env.FLAVOR);
+export default (async () => {
 	await firestore.collection('users').add({
 		name: 'John Doe',
 		age: 30,
 	});
-
-	return {
-		success: true,
-	};
-};
+}) satisfies ScriptFunction;