+++ src/index.ts
@@ -3,6 +3,6 @@ export * from './helpers/document-triggers';
 export * from './helpers/https';
 export * from './helpers/pubsub';
 export * from './helpers/storage';
-export * from './helpers/storagev2';
+export * from './helpers/storage-v2';
 export * from './helpers/https';
 export * from './types';